% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alpha2kmers}
\alias{alpha2kmers}
\title{Convert PWM model to kmers.}
\usage{
alpha2kmers(alpha, fasta)
}
\arguments{
\item{alpha}{PWM model to convert.}

\item{fasta}{Dataset of sequences.}
}
\value{
The best kmers from dataset with respect to alpha.
}
\description{
Convert PWM model to kmers.
}
