% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{char2int}
\alias{char2int}
\title{Converts char nucleotide A,C,G,T in int 0,1,2,3.}
\arguments{
\item{c}{char to convert for.}
}
\value{
The correspondent int or throws a invalid character.
}
\description{
Converts char nucleotide A,C,G,T in int 0,1,2,3.
}
