% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeDKL}
\alias{computeDKL}
\title{Compute DKL between alpha and beta models.}
\usage{
computeDKL(alpha, beta, kmer)
}
\arguments{
\item{alpha}{PWM model.}

\item{beta}{0-order Markov model.}

\item{kmer}{Target sequence.}
}
\value{
Kullback-Leibler divergence between alpha and beta models.
}
\description{
Compute DKL between alpha and beta models.
}
