% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeDKLU}
\alias{computeDKLU}
\title{Compute DKL between alpha and uniform models.}
\usage{
computeDKLU(alpha, kmer)
}
\arguments{
\item{alpha}{PWM model.}

\item{kmer}{Target sequence.}
}
\value{
Kullback-Leibler divergence between alpha and uniform models.
}
\description{
Compute DKL between alpha and uniform models.
}
