% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeIC}
\alias{computeIC}
\title{Compute information content.}
\usage{
computeIC(alpha, beta)
}
\arguments{
\item{alpha}{PWM model.}

\item{beta}{Markov 0-order model.}
}
\value{
Information content score.
}
\description{
Compute information content.
}
