% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeICU}
\alias{computeICU}
\title{Compute information content from a uniform distribution.}
\usage{
computeICU(alpha)
}
\arguments{
\item{alpha}{PWM model.}
}
\value{
Information content score.
}
\description{
Compute information content from a uniform distribution.
}
