% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corr_freq}
\alias{corr_freq}
\title{Computes the correlation frequency from correlation string.}
\usage{
corr_freq(correlation_str)
}
\arguments{
\item{correlation_str}{Correlation string computed from corr.}
}
\value{
Frenquency correlation.
}
\description{
Computes the correlation frequency from correlation string.
}
