% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createMarkovChain}
\alias{createMarkovChain}
\title{Create tau-order Markov Chain from the dataset.}
\usage{
createMarkovChain(fasta, tau = 0L)
}
\arguments{
\item{fasta}{The dataset os sequences.}

\item{tau}{The order of markov chain.}
}
\value{
The Markov Model of tau-order.
}
\description{
Create tau-order Markov Chain from the dataset.
}
