% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_corr_freq}
\alias{fast_corr_freq}
\title{Computes the correlation frequency between two strings.}
\usage{
fast_corr_freq(a, b)
}
\arguments{
\item{a}{First string.}

\item{b}{Second string.}
}
\value{
Frenquency correlation.
}
\description{
Computes the correlation frequency between two strings.
}
