% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hamming_distance_parallel}
\alias{hamming_distance_parallel}
\title{Computes hamming distance for two strings.}
\arguments{
\item{str1}{The first string.}

\item{str2}{The second string.}
}
\value{
The hamming distance between str1 and str2.
}
\description{
Computes hamming distance for two strings.
}
