% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hard_update}
\alias{hard_update}
\title{Update the alpha model with all kmers in the sequence.}
\usage{
hard_update(alpha, seq, posteriori)
}
\arguments{
\item{alpha}{PWM model.}

\item{seq}{All kmers in the sequence.}

\item{posteriori}{All kmers posteriori distribution.}
}
\description{
Update the alpha model with all kmers in the sequence.
}
