% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hasConverged}
\alias{hasConverged}
\title{Check if EM is converged.}
\arguments{
\item{cutoff}{If change is less than cutoff, EM has converged.}
}
\value{
True or False dependent on whether the EM has converged or not.
}
\description{
Check if EM is converged.
}
