% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{index2kmer}
\alias{index2kmer}
\title{Convert a integer to corresponding kmer.}
\arguments{
\item{index}{Index to converting for.}

\item{k}{The size of kmer.}
}
\value{
The corresponding kmer.
}
\description{
Convert a integer to corresponding kmer.
}
