% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int2char}
\alias{int2char}
\title{Converts int 0,1,2,3 to nucleotide char A,C,G,T.}
\arguments{
\item{i}{int to convert for.}
}
\value{
The correpondent nucleotide character or throws a invalid integer.
}
\description{
Converts int 0,1,2,3 to nucleotide char A,C,G,T.
}
