% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmer2index}
\alias{kmer2index}
\title{Convert a kmer into a corresponding integer.}
\arguments{
\item{kmer}{Kmer to convert for.}
}
\value{
A corresponding index to kmer.
}
\description{
Convert a kmer into a corresponding integer.
}
