% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmers2alpha}
\alias{kmers2alpha}
\title{Convert kmers to PWM model.}
\usage{
kmers2alpha(kmers)
}
\arguments{
\item{kmers}{Kmers to convert.}
}
\value{
PWM model from kmers.
}
\description{
Convert kmers to PWM model.
}
