% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LL}
\alias{LL}
\title{Computes de Log-Likelihood of the data and parameters.}
\usage{
LL(fasta, alpha, beta)
}
\arguments{
\item{fasta}{Dataset of sequences.}

\item{alpha}{PWM motif model.}

\item{beta}{Markov backround model.}
}
\value{
The Log-Likelihood of all sequences and parameters.
}
\description{
Computes de Log-Likelihood of the data and parameters.
}
