% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{probSeqGivenAlpha}
\alias{probSeqGivenAlpha}
\title{Computes de probability of a sequence give the alpha model.}
\usage{
probSeqGivenAlpha(kmer, alpha)
}
\arguments{
\item{kmer}{Sequence with k size.}

\item{alpha}{PWM (Position Weight Matrix) model.}
}
\value{
The probability of sequence given the current PWM model.
}
\description{
Computes de probability of a sequence give the alpha model.
}
