% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{probSeqGivenAlphaLog}
\alias{probSeqGivenAlphaLog}
\title{Computes de log-probability of a sequence give the alpha model.}
\usage{
probSeqGivenAlphaLog(kmer, alpha)
}
\arguments{
\item{kmer}{Sequence with k size.}

\item{alpha}{PWM (Position Weight Matrix) model.}
}
\value{
The probability of sequence given the current PWM model.
}
\description{
Computes de log-probability of a sequence give the alpha model.
}
