% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{probSeqGivenBeta}
\alias{probSeqGivenBeta}
\alias{probSeqGivenBetaLog}
\title{Computes de probability of a sequence give the beta model.}
\usage{
probSeqGivenBeta(seq, beta)

probSeqGivenBetaLog(seq, beta)
}
\arguments{
\item{seq}{Sequence with k size.}

\item{beta}{Markov Chain.}
}
\value{
The probability of sequence given the Markov model.

The probability of sequence given the Markov model.
}
\description{
Computes de probability of a sequence give the beta model.

Computes de probability of a sequence give the beta model.
}
