% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{probSeqGivenPosLog}
\alias{probSeqGivenPosLog}
\title{Computes de log-probability of a sequence give the position.}
\usage{
probSeqGivenPosLog(seq, alpha, beta, pos)
}
\arguments{
\item{seq}{Sequence with k size.}

\item{alpha}{PWM (Position Weight Matrix) model.}

\item{beta}{Markov Chain.}

\item{pos}{Position to compute probabilities.}
}
\value{
The probability of sequence given the position.
}
\description{
Computes de log-probability of a sequence give the position.
}
