% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{score}
\alias{score}
\alias{computeSCORE}
\title{Compute score from PWM model.}
\usage{
computeSCORE(alpha)
}
\arguments{
\item{alpha}{PWM model.}
}
\value{
Simple score from PWM model.
}
\description{
Compute score from PWM model.
}
