% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soft_update}
\alias{soft_update}
\title{Update the alpha model with all kmers in the sequence.}
\usage{
soft_update(alpha, seq, posteriori)
}
\arguments{
\item{alpha}{PWM model.}

\item{seq}{All kmers in the sequence.}

\item{posteriori}{All kmers posteriori distribution.}
}
\description{
Update the alpha model with all kmers in the sequence.
}
