% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{step_size}
\alias{step_size}
\title{Compute the step-size for EM convergence.}
\arguments{
\item{convergence}{Convergence vector.}
}
\value{
The abs of difference between the 2 last positions of convercence vector.
}
\description{
Compute the step-size for EM convergence.
}
