% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{string_corr}
\alias{string_corr}
\alias{corr}
\title{This function computes the overlap string between two string a and b.}
\usage{
corr(a, b)
}
\arguments{
\item{a}{First string.}

\item{b}{Second string.}
}
\value{
The correlation string between both strings a and b.
}
\description{
This function computes the overlap string between two string a and b.
}
